packages/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
web3==6.11.1
requests==2.31.0
pytest==7.4.3
httpx==0.25.1 